select box

+++ app/src/main/java/pe/edu/ulima/pm20232/aulavirtual/models/Pokemon.kt
@@ -1,12 +1,12 @@
 package pe.edu.ulima.pm20232.aulavirtual.models
 
 data class Pokemon (
-    val id: Long,
+    val id: Int,
     val name: String,
     val number: Long,
     val weight: Double,
     val height: Double,
     val imageUrl: String,
-    val generationId: Long,
+    val generationId: Int,
     val generationName: String
 )
